+++ examples/clienttest/lib/constants/theme.dart
@@ -1,5 +1,0 @@
-import 'package:jaspr/jaspr.dart';
-
-// As your css styles are defined using just Dart, you can simply
-// use global variables or methods for common things like colors.
-const primaryColor = Color('#01589B');