+++ examples/clienttest/lib/jaspr_options.client.g.dart
@@ -1,9 +1,0 @@
-// dart format off
-// ignore_for_file: type=lint
-
-// GENERATED FILE, DO NOT MODIFY
-// Generated with jaspr_builder
-
-import 'package:jaspr/browser.dart';
-
-ClientOptions get defaultClientOptions => ClientOptions();