+++ examples/clienttest/lib/main.client.dart
@@ -1,9 +1,0 @@
-import 'package:jaspr/browser.dart';
-
-import 'jaspr_options.client.g.dart';
-
-void main() {
-  runApp(
-    ClientApp(options: defaultClientOptions),
-  );
-}