+++ packages/jaspr/lib/src/server/run_app.dart
@@ -4,8 +4,8 @@ import 'dart:typed_data';
 
 import 'package:shelf/shelf.dart';
 
-import 'options.dart';
 import '../framework/framework.dart';
+import 'options.dart';
 import 'render_functions.dart';
 import 'server_app.dart';
 import 'server_handler.dart';