+++ packages/jaspr_cli/tool/scaffold/__brick__/lib/client#main.client.dart
@@ -1,6 +1,7 @@
-// The entrypoint for the **client** app.
-//
-// This file is compiled to javascript and executed in the browser.
+/// The entrypoint for the **client** app.
+///
+/// This file is compiled to javascript and executed on the client when loading the page.
+library;
 
 // Client-specific Jaspr import.
 import 'package:jaspr/browser.dart';